added loss selection

+++ optimization/shared/fed_avg_schedule_loss_selection.py
@@ -506,7 +506,7 @@ def build_fed_avg_process(
     zero = tf.zeros(shape=[0,1] , dtype=tf.float32)
     # list_type = tff.SequenceType( tff.TensorType(dtype=tf.float32))
     list_type = tff.TensorType(dtype = tf.float32)
-    one_value = tff.TensorType()
+    one_value = tff.TensorType(dtype = tf.float32)
     @computations.tf_computation(list_type, tf.float32)
     def accumulate(u,t):
       t = tf.reshape(t, shape=[1,1])
